refacto: split loaders into separate modules
